fixup! adding export_countries to Company API
