--- datahub/interaction/constants.py
@@ -1,1 +1,0 @@
-INTERACTION_ADD_COUNTRIES = 'interaction-add-countries'
